Follow personal parenthesis code convention in test files
If the intent of a method call is to return a value, and the caller of
the method or the variable will need to store/obtain the value, use
parenthesis. If it's an imperative method call (i.e., a command), don't
use parenthesis.

Using parenthesis makes it clear that the value is being returned and
is meaningful.