Improve application performance
The mechanize gem takes a long time to load, and was the main bottleneck
in my profiling tests. I'm requiring mechanize only when needed. The
tests are now faster, as well as general performance. It can be further
improved, but it's fine for now.

Also, everything is now required manually.